Partition things by the hours
This automatically fixes the offset. If we start at hour 5, the first
6 hour set will simply contain hour 5 by virtue of the hour
being 5. No need to fudge with dynamic offsets passed in from
the outside.

A convenience offset will still be needed, in case we want to have "morning"
start at (say) 7:00 instead of 6:00. But do we want that at all? Or have
"morning" be from 3:00 up to 9:00, and then noon from 9:00 to 15:00?